requirements.txt
appdirs==1.4.0
CacheControl==0.10.7
Flask==0.10.1
PyYAML==3.11
SQLAlchemy==0.9.8
ansicolors==1.0.2
click
github3.py
gunicorn==18.0
iso8601
lockfile==0.9.1
more_itertools
psycopg2==2.5.4
pylint
python-dateutil
requests==2.5.0
statistics
uritemplate.py==0.2.0
urlobject
wsgiref==0.1.2
jira
scrapy
path.py
gitpython
python-dotenv
pytest


-e .
